Move helper functions to utils.py.
