+++ test/test_test_function.py
@@ -1,10 +1,9 @@
-import json
 import unittest
 
 from jsonschema import ValidationError
 
 from rest_tester.test_function import TestFunction
-from rest_tester.test_info import TestInfo
+from test.utils import load_test_info
 
 
 class TestTestFunction(unittest.TestCase):
@@ -39,7 +38,7 @@ class TestTestFunction(unittest.TestCase):
 
     def test_make_test_status_code(self):
         json_file = './resources/status_code_success.json'
-        url, params, timeout, test_cases = self.load_test_info(json_file)
+        url, params, timeout, test_cases = load_test_info(json_file)
 
         test_key = 'statusCode'
 
@@ -47,7 +46,7 @@ class TestTestFunction(unittest.TestCase):
 
     def test_make_test_status_code_fail(self):
         json_file = './resources/status_code_fail.json'
-        url, params, timeout, test_cases = self.load_test_info(json_file)
+        url, params, timeout, test_cases = load_test_info(json_file)
 
         test_key = 'statusCode'
 
@@ -55,7 +54,7 @@ class TestTestFunction(unittest.TestCase):
 
     def test_make_test_json_schema(self):
         json_file = './resources/json_schema_success.json'
-        url, params, timeout, test_cases = self.load_test_info(json_file)
+        url, params, timeout, test_cases = load_test_info(json_file)
 
         test_key = 'jsonSchema'
 
@@ -63,7 +62,7 @@ class TestTestFunction(unittest.TestCase):
 
     def test_make_test_json_schema_fail(self):
         json_file = './resources/json_schema_fail.json'
-        url, params, timeout, test_cases = self.load_test_info(json_file)
+        url, params, timeout, test_cases = load_test_info(json_file)
 
         test_key = 'jsonSchema'
 
@@ -71,29 +70,25 @@ class TestTestFunction(unittest.TestCase):
 
     def test_make_test_unsupported_case(self):
         json_file = './resources/unsupported_case.json'
-        url, params, timeout, test_cases = self.load_test_info(json_file)
+        url, params, timeout, test_cases = load_test_info(json_file)
 
         test_key = 'unsupportedCase'
 
         self.run_fail_function(url, params, timeout, test_key, test_cases[test_key], AssertionError)
 
     def run_success_function(self, url, params, timeout, test_key, test_case):
+        """Helper function for successful run"""
         test_function = TestFunction.make(url, params, timeout, test_key, test_case)
         test_function(self)
 
     def run_fail_function(self, url, params, timeout, test_key, test_case, error):
+        """Helper function for unsuccessful run; should throw given error"""
         test_function = TestFunction.make(url, params, timeout, test_key, test_case)
         try:
             test_function(self)
             self.fail('Should throw AssertionError!')
         except error:
             pass
-
-    @staticmethod
-    def load_test_info(json_file):
-        with open(json_file, 'r') as json_file:
-            json_data = json.load(json_file)
-            return TestInfo.read(json_data)
 
 
 if __name__ == '__main__':
